Update and improve plugin
