--- .distignore
@@ -1,5 +1,0 @@
-/.wordpress-org
-/.git
-/.github
-.distignore
-.gitignore
