+++ .github/workflows/deploy.yml
@@ -8,11 +8,16 @@ jobs:
     name: New tag
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@master
+    - name: Checkout code
+      uses: actions/checkout@master
+    - name: Set up PHP environment
+      uses: shivammathur/setup-php@v2
+      with:
+        php-version: '7.4'
     - name: Build Plugin
       run: make dist
     - name: WordPress Plugin Deploy
-      uses: woocart/action-wordpress-plugin-deploy@master
+      uses: akshitsethi/action-wordpress-plugin-deploy@master
       env:
         SVN_PASSWORD: ${{ secrets.SVN_PASSWORD }}
         SVN_USERNAME: ${{ secrets.SVN_USERNAME }}
