+++ .gitignore
@@ -12,3 +12,5 @@ reports
 clover.xml
 src/vendor
 dist
+node_modules
+.phpunit.result.cache
