+++ composer.json
@@ -1,30 +1,27 @@
 {
   "require": {
-    "php": ">=7.2.0",
+    "php": ">=7.3.0",
     "ext-gd": "*",
     "ext-json": "*"
   },
   "require-dev": {
-    "dealerdirect/phpcodesniffer-composer-installer": "*",
-    "phpunit/phpunit": "6.5.9",
-    "10up/wp_mock": "0.3.0",
-    "antecedent/patchwork": "*",
-    "php-coveralls/php-coveralls": "^2.1",
-    "squizlabs/php_codesniffer": "*",
-    "wp-coding-standards/wpcs": "*",
-    "rregeer/phpunit-coverage-check": "^0.1.6",
-    "phpunit/phpcov": "*",
-    "phpro/grumphp": "^0.14.2",
-    "sensiolabs/security-checker": "^5.0",
-    "maglnet/composer-require-checker": "^1.1",
-    "wp-cli/wp-cli": "^2.0"
+    "dealerdirect/phpcodesniffer-composer-installer": "^0.7.1",
+    "phpunit/phpunit": "^9.5.0",
+    "10up/wp_mock": "^0.4.2",
+    "antecedent/patchwork": "^2.1.12",
+    "php-coveralls/php-coveralls": "^2.4.3",
+    "squizlabs/php_codesniffer": "^3.5.8",
+    "wp-coding-standards/wpcs": "^2.3.0",
+    "rregeer/phpunit-coverage-check": "^0.3.1",
+    "phpunit/phpcov": "^8.2.0",
+    "phpro/grumphp": "^1.3.0"
   },
   "config": {
     "bin-dir": "bin/"
   },
   "autoload": {
     "psr-4": {
-      "Niteo\\WooCart\\WooDash\\": "classes"
+      "Niteo\\WooCart\\WooDash\\": "src/inc"
     }
   }
 }
