--- src/classes/class-config.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Niteo\WooCart\WooDash {
-
-	/**
-	 * Configuration options for the plugin.
-	 *
-	 * @package    Niteo\WooCart\WooDash
-	 * @since      1.0.0
-	 */
-	class Config {
-
-		const DB_OPTION      = 'woodash_options';
-		const PREFIX         = 'woodash_';
-		const DEFAULT_STATUS = 'woodash';
-
-	}
-
-}
