+++ src/composer.json
@@ -1,5 +1,5 @@
 {
-  "name": "WooDash",
+  "name": "woocart/woodash",
   "description": "",
   "keywords": [
     "WordPress",
@@ -8,11 +8,11 @@
   ],
   "type": "wordpress-plugin",
   "require": {
-    "php": ">=5.6.0"
+    "php": ">=7.3.0"
   },
   "autoload": {
     "psr-4": {
-      "Niteo\\WooCart\\WooDash\\": "classes"
+      "Niteo\\WooCart\\WooDash\\": "inc"
     }
   }
 }
