+++ src/inc/Config.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * Configuration file for the plugin.
+ */
+
+namespace Niteo\WooCart\WooDash;
+
+/**
+ * Configuration options for the plugin.
+ *
+ * @package Niteo\WooCart\WooDash
+ */
+class Config {
+
+	/**
+	 * @var string
+	 */
+	const DB_OPTION = 'woodash_options';
+
+	/**
+	 * @var string
+	 */
+	const PREFIX = 'woodash_';
+
+	/**
+	 * @var string
+	 */
+	const DEFAULT_STATUS = 'woodash';
+
+	/**
+	 * @var float
+	 */
+	const MIN_WC_VERSION = '3.1.0';
+
+}
