+++ src/readme.txt
@@ -2,9 +2,9 @@
 Contributors: woocart
 Tags: woocart, woocommerce, woodash, woo dashboard
 Requires at least: 4.6
-Tested up to: 5.3.2
-Stable tag: 1.2.0
-Requires PHP: 5.6
+Tested up to: 5.7.0
+Stable tag: 1.3.0
+Requires PHP: 7.3
 License: GPLv3 or later
 License URI: https://www.gnu.org/licenses/gpl-3.0.html
 
@@ -48,12 +48,18 @@ https://github.com/woocart/woodash
 
 
 == Changelog ==
+
+= v1.3.0 =
+* Compatible with WP v5.7
+* Code refactoring
+* Support for PHP 8
+
 = v1.2.0 =
-- Shows admin notice on plugin activation
-- Taxes in the admin menu are hidden if the option is not enabled in WooCommerce
+* Shows admin notice on plugin activation
+* Taxes in the admin menu are hidden if the option is not enabled in WooCommerce
 
 = v1.1.3 =
-- Fixed conflict with WooCommerce Admin plugin
+* Fixed conflict with WooCommerce Admin plugin
 
 == Upgrade Notice ==
