+++ tests/MenuTest.php
@@ -1,58 +1,67 @@
 <?php
+/**
+ * Unit tests for `Menu` class.
+ */
 
+namespace Niteo\WooCart\WooDash\Tests;
 
 use Niteo\WooCart\WooDash\Menu;
 use PHPUnit\Framework\TestCase;
 
+/**
+ * Tests Menu class functions in isolation.
+ *
+ * @package Niteo\WooCart\WooDash
+ * @coversDefaultClass \Niteo\WooCart\WooDash\Menu
+ */
 class MenuTest extends TestCase {
 
-
-	function setUp() {
+	function setUp() : void {
+		\WP_Mock::setUsePatchwork( true );
 		\WP_Mock::setUp();
 	}
 
-	function tearDown() {
+	function tearDown() : void {
 		$this->addToAssertionCount(
 			\Mockery::getContainer()->mockery_getExpectationCount()
 		);
+
 		\WP_Mock::tearDown();
-		\Mockery::close();
 	}
 
-
 	/**
-	 * @covers \Niteo\WooCart\WooDash\Menu::__construct
+	 * @covers ::__construct
 	 */
 	public function testConstructor() {
 		$menu = new Menu();
 
 		$this->assertEquals(
-			[
-				'orders'    => [
+			array(
+				'orders'    => array(
 					'name'     => 'Orders',
 					'link'     => 'edit.php?post_type=shop_order',
 					'priority' => 101,
 					'icon'     => 'dashicons-heart',
-				],
-				'stock'     => [
+				),
+				'stock'     => array(
 					'name'     => 'Stock',
 					'link'     => 'admin.php?page=wc-reports&tab=stock',
 					'priority' => 102,
 					'icon'     => 'dashicons-archive',
-				],
-				'customers' => [
+				),
+				'customers' => array(
 					'name'     => 'Customers',
 					'link'     => 'admin.php?page=wc-reports&tab=customers&report=customer_list',
 					'priority' => 103,
 					'icon'     => 'dashicons-groups',
-				],
-				'reports'   => [
+				),
+				'reports'   => array(
 					'name'     => 'All Reports',
 					'link'     => 'admin.php?page=wc-reports',
 					'priority' => 105,
 					'icon'     => 'dashicons-chart-area',
-				],
-			],
+				),
+			),
 			$menu->__construct()
 		);
 	}
