+++ tests/autoload.php
@@ -2,9 +2,8 @@
 
 $root_dir = dirname( dirname( __FILE__ ) );
 
-require_once "$root_dir/vendor/autoload.php";
-require_once "$root_dir/src/woodash.php";
+require_once "{$root_dir}/vendor/autoload.php";
+require_once "{$root_dir}/vendor/antecedent/patchwork/Patchwork.php";
 
 WP_Mock::setUsePatchwork( true );
 WP_Mock::bootstrap();
-
